requirements.txt
dj-database-url==0.5.0
Django==2.0.2
django-ckeditor==5.9.0
django-heroku==0.3.1
django-js-asset==1.2.2
freeze==3.0
gunicorn==20.0.4
mysqlclient==2.0.1
Pillow==7.2.0
psycopg2==2.8.6
pytz==2020.1
six==1.15.0
whitenoise==5.2.0
